% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boschek.test.R
\docType{data}
\name{d1}
\alias{d1}
\title{d1}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{V1}{ratings by rater 1}
  \item{V2}{ratings by rater 2}
  \item{V3}{ratings by rater 3}
  \item{V4}{ratings by rater 4}
}}
\usage{
d1
}
\description{
test dataset for ilustration
}
\keyword{datasets}
