% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boschek.test.R
\docType{data}
\name{d2}
\alias{d2}
\title{d2}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{V1}{rating by rater 1}
  \item{V2}{rating by rater 2}
  \item{V3}{rating by rater 3}
  \item{frequency}{number of rating with the particular combination of ratings}
}}
\usage{
d2
}
\description{
test dataset for ilustration
}
\keyword{datasets}
